+++ crates/cli-tools/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Minor
 
+- Increase the default connection timeout from 1 to 5 seconds
+- Add `action::PlatformUpdate` for platform update
 - Add `action::PlatformList` to list connected platforms
 - Add `action::ConnectionOptions` for commands that need a platform connection
 - Change the behavior of `fs::copy_if_changed()` to keep an original source
@@ -18,4 +20,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 3 -->
+<!-- Increment to skip CHANGELOG.md test: 4 -->