+++ crates/cli/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Rename `--serial` to `--protocol` with more support
 - Move `--serial` and `--timeout` to commands that need them
 
+### Minor
+
+- Add `platform-update-{metadata,transfer}` for platform update
+
 ### Patch
 
 - Check for bad command-line configuration
@@ -36,4 +40,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 6 -->
+<!-- Increment to skip CHANGELOG.md test: 7 -->