+++ crates/protocol/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Minor
 
+- Add a `Service::NAME` constant with `host` feature
+- Add `PlatformUpdate{Metadata,Transfer}` calls and `transfer` module for platform updates
 - Add a `Connection` abstraction with `host` feature
 
 ### Patch
@@ -12,4 +14,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 3 -->
+<!-- Increment to skip CHANGELOG.md test: 4 -->